@@ -1,7 +1,7 @@
 import { CommonModule } from '@angular/common';
 import { AgGridModule } from 'ag-grid-angular';
 
-import { Component, OnInit, Output, EventEmitter, inject } from '@angular/core';
+import { Component, OnInit, Output, EventEmitter, inject, output } from '@angular/core';
 
 import { AppAlarmService } from 'src/app/core/service/app-alarm.service';
 import { AggridFunction } from 'src/app/core/grid/aggrid-function';
@@ -34,9 +34,9 @@ export class BizCodeTypeGridComponent extends AggridFunction implements OnInit {
 
   _list: BizCodeType[] = [];
 
-  @Output() rowClickedEvent = new EventEmitter();
-  @Output() rowDoubleClickedEvent = new EventEmitter();
-  @Output() editButtonClickedEvent = new EventEmitter();
+  rowClicked = output<any>();
+  rowDoubleClicked = output<any>();
+  editButtonClicked = output<any>();
 
   private service = inject(BizCodeTypeService);
   private appAlarmService = inject(AppAlarmService);
@@ -92,15 +92,15 @@ export class BizCodeTypeGridComponent extends AggridFunction implements OnInit {
 
   rowClickedFunc(event: any): void {
     const selectedRows = this.gridApi.getSelectedRows();
-    this.rowClickedEvent.emit(selectedRows[0]);
+    this.rowClicked.emit(selectedRows[0]);
   }
 
   rowDoubleClickedFunc(event: any): void {
-    this.rowDoubleClickedEvent.emit(event.data);
+    this.rowDoubleClicked.emit(event.data);
   }
 
   onEditButtonClick(e: any): void {
-    this.editButtonClickedEvent.emit(e.rowData);
+    this.editButtonClicked.emit(e.rowData);
   }
 
 }
