@@ -1,7 +1,7 @@
 import { CommonModule } from '@angular/common';
 import { AgGridModule } from 'ag-grid-angular';
 
-import { Component, OnInit, Output, EventEmitter, Input, inject } from '@angular/core';
+import { Component, OnInit, Output, EventEmitter, Input, inject, output } from '@angular/core';
 
 import { AppAlarmService } from 'src/app/core/service/app-alarm.service';
 import { AggridFunction } from 'src/app/core/grid/aggrid-function';
@@ -36,9 +36,10 @@ export class MenuGridComponent extends AggridFunction implements OnInit {
 
   menuList: Menu[] = [];
 
-  @Output() rowSelected = new EventEmitter();
-  @Output() editButtonClicked = new EventEmitter();
-  @Output() rowDoubleClicked = new EventEmitter();
+  rowClicked = output<any>();
+  rowDoubleClicked = output<any>();
+  editButtonClicked = output<any>();
+
   @Input() menuGroupCode: string = '';
 
   private menuService = inject(MenuService);
@@ -105,7 +106,7 @@ export class MenuGridComponent extends AggridFunction implements OnInit {
   selectionChanged(event: any) {
     const selectedRows = this.gridApi.getSelectedRows();
 
-    this.rowSelected.emit(selectedRows[0]);
+    this.rowClicked.emit(selectedRows[0]);
   }
 
   rowDbClicked(event: any) {
