@@ -1,7 +1,7 @@
 import { CommonModule } from '@angular/common';
 import { AgGridModule } from 'ag-grid-angular';
 
-import { Component, OnInit, Output, EventEmitter, inject } from '@angular/core';
+import { Component, OnInit, Output, EventEmitter, inject, output } from '@angular/core';
 
 import { AppAlarmService } from 'src/app/core/service/app-alarm.service';
 import { AggridFunction } from 'src/app/core/grid/aggrid-function';
@@ -37,9 +37,9 @@ export class MenuGroupGridComponent extends AggridFunction implements OnInit {
 
   menuGroupList: MenuGroup[] = [];
 
-  @Output() rowSelected = new EventEmitter();
-  @Output() editButtonClicked = new EventEmitter();
-  @Output() rowDoubleClicked = new EventEmitter();
+  rowClicked = output<any>();
+  rowDoubleClicked = output<any>();
+  editButtonClicked = output<any>();
 
   private menuService = inject(MenuService);
   private appAlarmService = inject(AppAlarmService);
@@ -121,7 +121,7 @@ export class MenuGroupGridComponent extends AggridFunction implements OnInit {
   selectionChanged(event: any) {
     const selectedRows = this.gridApi.getSelectedRows();
 
-    this.rowSelected.emit(selectedRows[0]);
+    this.rowClicked.emit(selectedRows[0]);
   }
 
   rowDbClicked(event: any) {
