@@ -1,7 +1,7 @@
 import { CommonModule } from '@angular/common';
 import { AgGridModule } from 'ag-grid-angular';
 
-import { Component, OnInit, Output, EventEmitter, inject } from '@angular/core';
+import { Component, OnInit, Output, EventEmitter, inject, output } from '@angular/core';
 
 import { ResponseList } from 'src/app/core/model/response-list';
 import { AggridFunction } from 'src/app/core/grid/aggrid-function';
@@ -37,9 +37,9 @@ export class TermGridComponent extends AggridFunction implements OnInit {
 
   termList: Term[] = [];
 
-  @Output() rowClickedEvent = new EventEmitter();
-  @Output() rowDoubleClickedEvent = new EventEmitter();
-  @Output() editButtonClickedEvent = new EventEmitter();
+  rowClicked = output<any>();
+  rowDoubleClicked = output<any>();
+  editButtonClicked = output<any>();;
 
   private termService = inject(TermService);
   private appAlarmService = inject(AppAlarmService);
@@ -99,15 +99,15 @@ export class TermGridComponent extends AggridFunction implements OnInit {
 
   rowClickedFunc(event: any) {
     const selectedRows = this.gridApi.getSelectedRows();
-    this.rowClickedEvent.emit(selectedRows[0]);
+    this.rowClicked.emit(selectedRows[0]);
   }
 
   onEditButtonClick(event: any) {
-    this.editButtonClickedEvent.emit(event.rowData);
+    this.editButtonClicked.emit(event.rowData);
   }
 
   rowDbClickedFunc(event: any) {
-    this.rowDoubleClickedEvent.emit(event.data);
+    this.rowDoubleClicked.emit(event.data);
   }
 
 }
