@@ -3,7 +3,7 @@ import { FormsModule } from '@angular/forms';
 import { NzTreeComponent, NzTreeModule, NzTreeNode } from 'ng-zorro-antd/tree';
 import { NzFormatEmitEvent } from 'ng-zorro-antd/tree';
 
-import { Component, OnInit, Output, EventEmitter, Input, inject, viewChild, input } from '@angular/core';
+import { Component, OnInit, Output, EventEmitter, Input, inject, viewChild, input, output } from '@angular/core';
 import { ResponseList } from 'src/app/core/model/response-list';
 
 import { MenuRoleHierarchy } from '../menu/menu-role-hierarchy.model';
@@ -57,10 +57,11 @@ export class MenuRoleTreeComponent implements OnInit {
   defaultCheckedKeys = ['1'];
 
   @Input() searchValue = '';
-  @Output() itemSelected = new EventEmitter();
 
   menuGroupCode = input<string>();
   roleCode = input<string>();
+
+  itemSelected = output<any>();
 
   private menuService = inject(MenuRoleService);
 
